+++ src/main/java/com/createver/server/global/client/OpenAiApiClient.java
@@ -18,7 +18,7 @@ import org.springframework.web.client.RestTemplate;
 
 @Service
 @RequiredArgsConstructor
-public class OpenAiService {
+public class OpenAiApiClient {
 
     private final RestTemplate restTemplate;
