+++ src/test/java/com/createver/server/domain/image/service/avatar/ImageAvatarServiceTest.java
@@ -3,7 +3,7 @@ package com.createver.server.domain.image.service;
 import com.createver.server.domain.image.dto.request.AvatarPromptRequest;
 import com.createver.server.domain.image.entity.ImageAvatar;
 import com.createver.server.domain.image.repository.avatar.ImageAvatarRepository;
-import com.createver.server.domain.image.service.avatar.ImageAvatarService;
+import com.createver.server.domain.image.service.avatar.ImageAvatarGenerationService;
 import com.createver.server.domain.member.entity.Member;
 import com.createver.server.domain.member.repository.MemberRepository;
 import org.junit.jupiter.api.DisplayName;
@@ -29,7 +29,7 @@ import static org.mockito.Mockito.*;
 class ImageAvatarServiceTest {
 
     @InjectMocks
-    private ImageAvatarService imageAvatarService;
+    private ImageAvatarGenerationService imageAvatarService;
 
     @Mock
     private RestTemplate restTemplate;
