+++ src/test/java/com/createver/server/domain/image/service/avatar/ImageAvatarSseServiceTest.java
